pyproject.toml
[project]
name = "unofficial-pecron-api"
version = "0.1.0"
description = "Unofficial Python API client for Pecron portable power stations"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Jesse Sightler", email = "jesse.sightler@gmail.com" }
]
requires-python = ">=3.11"
keywords = ["pecron", "power-station", "iot", "home-assistant", "quectel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.28",
    "pycryptodome>=3.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
pecron = "unofficial_pecron_api.cli:main"

[project.urls]
Homepage = "https://github.com/jsightler/unofficial-pecron-api"
Repository = "https://github.com/jsightler/unofficial-pecron-api"
Issues = "https://github.com/jsightler/unofficial-pecron-api/issues"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
